requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
yfinance>=0.2.28
pandas>=2.0.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
boto3>=1.34.0
pydantic>=2.0.0
scikit-learn>=1.3.0